refractor: rename AppStore.js to store.js
Rename as store to follow best practices
